fix a lot of integer type warnings

+++ src/lib/neural.h
@@ -4,12 +4,12 @@
 
 struct Neuron {
   float* input_weights;
-  int input_count;
+  size_t input_count;
 };
 
 struct NeuralNet {
-  int layers;
-  int neurons_per_layer;
+  size_t layers;
+  size_t neurons_per_layer;
 
   struct Neuron** neurons;
   struct Neuron output_neuron;
@@ -18,59 +18,59 @@ struct NeuralNet {
 struct Net_Training_Settings {
   float learning_rate;
   float acceptable_error;
-  int max_changes_at_once;
-  int num_threads;
-  int iterations_per_thread;
+  size_t max_changes_at_once;
+  size_t num_threads;
+  size_t iterations_per_thread;
 };
 
 struct Net_Training_Item {
   float* inputs;
-  int input_count;
+  size_t input_count;
 
   float correct_output;
 };
 
 struct NeuralNet_Weight_Adjustment {
   float amount;
-  int weight;
+  size_t weight;
 
   struct Neuron* neuron;
 
   int applied;
 };
 
-int get_rand();
+long int get_rand();
 float random_sign();
 float random_float();
 float sign_of(float x);
 
-void apply_weight_adjustments(struct NeuralNet_Weight_Adjustment* adjustments, int count);
-void undo_weight_adjustments(struct NeuralNet_Weight_Adjustment* adjustments, int count);
+void apply_weight_adjustments(struct NeuralNet_Weight_Adjustment* adjustments, size_t count);
+void undo_weight_adjustments(struct NeuralNet_Weight_Adjustment* adjustments, size_t count);
 
-struct NeuralNet_Weight_Adjustment generate_weight_adjustment(struct Neuron* neuron, int weight, float amount);
+struct NeuralNet_Weight_Adjustment generate_weight_adjustment(struct Neuron* neuron, size_t weight, float amount);
 struct NeuralNet_Weight_Adjustment generate_random_weight_adjustment(struct NeuralNet* net, struct Net_Training_Settings settings);
 
 void free_neuron(struct Neuron* neuron);
 void free_neural_net(struct NeuralNet* neural_net);
 
-struct Neuron create_neuron(int input_count);
+struct Neuron create_neuron(size_t input_count);
 int copy_neuron(struct Neuron* from, struct Neuron* to);
 struct Neuron duplicate_neuron(struct Neuron dup);
 
-struct NeuralNet create_neural_net(int layers, int neurons_per_layer);
+struct NeuralNet create_neural_net(size_t layers, size_t neurons_per_layer);
 int copy_neural_net(struct NeuralNet* from, struct NeuralNet* to);
 struct NeuralNet duplicate_neural_net(struct NeuralNet dup);
 
 float activation_tanh(float output);
 
-float neuron_evaluate(struct Neuron neuron, float* input, int input_count);
-void neural_layer_evaluate(struct Neuron* layer, int layer_count, float* input, int input_count, float* layer_outputs);
-float neural_evaluate(struct NeuralNet* net, float* input, int input_count);
-float neural_error(struct NeuralNet* net, float* input, int input_count, float correct_output);
+float neuron_evaluate(struct Neuron neuron, float* input, size_t input_count);
+void neural_layer_evaluate(struct Neuron* layer, size_t layer_count, float* input, size_t input_count, float* layer_outputs);
+float neural_evaluate(struct NeuralNet* net, float* input, size_t input_count);
+float neural_error(struct NeuralNet* net, float* input, size_t input_count, float correct_output);
 float neural_overall_error(struct NeuralNet* net, struct Net_Training_Item* items, int count);
 
 void neural_net_test_random_adjust(struct NeuralNet* net, struct Net_Training_Item* items, int item_count,
-                                   struct NeuralNet_Weight_Adjustment* adjustments, int adj_count);
+                                   struct NeuralNet_Weight_Adjustment* adjustments, size_t adj_count);
 
 int neural_train(struct NeuralNet* net, struct Net_Training_Settings settings, struct Net_Training_Item* items,
                  int item_count);
